extract base integration rule
