+++ src/test/java/com/rbkmoney/columbus/GeoServiceTest.java
@@ -2,10 +2,6 @@ package com.rbkmoney.columbus;
 
 import com.maxmind.geoip2.exception.GeoIp2Exception;
 import com.maxmind.geoip2.model.CityResponse;
-import com.palantir.docker.compose.DockerComposeRule;
-import com.palantir.docker.compose.connection.DockerPort;
-import com.palantir.docker.compose.connection.waiting.HealthChecks;
-import com.palantir.docker.compose.logging.FileLogCollector;
 import com.rbkmoney.columbus.dao.CityLocationsDao;
 import com.rbkmoney.columbus.dao.GeoIpDao;
 import com.rbkmoney.columbus.model.CityLocation;
@@ -17,22 +13,17 @@ import com.rbkmoney.damsel.base.InvalidRequest;
 import com.rbkmoney.damsel.geo_ip.GeoIDInfo;
 import com.rbkmoney.damsel.geo_ip.LocationInfo;
 import com.rbkmoney.damsel.geo_ip.geo_ipConstants;
-import com.zaxxer.hikari.HikariDataSource;
 import org.apache.thrift.TException;
 import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import javax.sql.DataSource;
-import java.io.File;
 import java.io.IOException;
 import java.util.*;
 
@@ -43,31 +34,11 @@ import static org.springframework.boot.test.context.SpringBootTest.WebEnvironmen
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = RANDOM_PORT)
 @TestPropertySource(locations="classpath:test.properties")
+@Import(IntegrationBaseRule.class)
 public class GeoServiceTest {
 
-    @TestConfiguration
-    static class Config {
-        @Bean
-        public DataSource dataSource() {
-            DockerPort postgres = docker.containers().container("postgres").port(5432);
-
-            return DataSourceBuilder
-                    .create()
-                    .type(HikariDataSource.class)
-                    .username("postgres")
-                    .password("postgres")
-                    .driverClassName("org.postgresql.Driver")
-                    .url(postgres.inFormat("jdbc:postgresql://$HOST:$EXTERNAL_PORT/postgres"))
-                    .build();
-        }
-    }
-
     @ClassRule
-    public static DockerComposeRule docker = DockerComposeRule.builder()
-            .file("src/test/resources/docker-compose.yml")
-            .logCollector(new FileLogCollector(new File("target/pglog")))
-            .waitingForService("postgres", HealthChecks.toHaveAllPortsOpen())
-            .build();
+    public static IntegrationBaseRule rule = new IntegrationBaseRule();
 
     public static final Map<String, String> IP_TO_CITY = new HashMap<>();
     public static final String IP_MOSCOW = "94.159.54.234";
